UPSTREAM: drivers/bus: arm-cci: function to enable CCI ports from early boot code
This provides cci_enable_port_for_self().  This is the counterpart to
cci_disable_port_by_cpu(self).

This is meant to be called from the MCPM machine specific power_up_setup
callback code when the appropriate affinity level needs to be initialized.
The code therefore has to be position independent as the MMU is still off
and it cannot rely on any stack space.

Signed-off-by: Nicolas Pitre <nico@linaro.org>
Reviewed-by: Dave Martin <dave.martin@linaro.org>
(cherry picked from linux-next commit
	62158f817ad3e0368480b8e9480be34bffd8c74a)

BUG=chrome-os-partner:19820
TEST=Boot on peach-pit.  Doesn't do anything since we're not using
this driver yet.

Change-Id: Ia11b4e8ec994d9974f8d2b14ef71037d74d6ea59
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59384