iommu/exynos: do not check MMU_CTRL during sysmmu enable/resume
On exynos5420, the reset value of MMU_CTRL of FIMD's sysmmu is not zero. This
causes the driver to hit a BUG in __sysmmu_enable_nocount. To avoid
this, we should not check the MMU_CTRL value during sysmmu_enable and
sysmmu_resume but only during sysmmu_runtime_resume. This patch adds
a check based on an init flag parameter.

Note: Currently we don't hit the BUG during bootup because this register
is set to zero in uboot. But we hit the BUG during system resume. Though this is
only for exynos5420's fimd sysmmu, there is harm in not checking for all
sysmmus.

BUG=chrome-os-partner:19007
TEST=Boot kernel on smdk5420 and run s2r. Boot on Snow and run s2r
Note: s2r on 5420 tested using rtcwake.

Change-Id: I0358aea346a1eedbed9518f2fcd639fb7b9bd4ee
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/56582
Tested-by: Prathyush Kalashwaram <prathyush@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>