UPSTREAM: rtc: add ability to push out an existing wakealarm using sysfs
This adds the ability for the rtc sysfs code to handle += characters at
the beginning of a wakealarm setting string.  This will allow the user to
attempt to push out an existing wakealarm by a provided amount.

In the case that the += characters are provided but the alarm is not
active -EINVAL is returned.

his is useful, at least for my purposes in suspend/resume testing.  The
basic test goes something like:

1. Set a wake alarm from userspace 5 seconds in the future

2. Start the suspend process (echo mem > /sys/power/state)

3. After ~2.5 seconds if userspace is still running (using another
   thread to check this), move the wake alarm 5 more seconds

If the "move" involves an unset of the wakealarm then there's a period
   of time where the system is midway through suspending but has no wake
   alarm.  It will get stuck.

We'd rather not remove the "move" since the idea is to avoid a
cancelled suspend when the alarm fires _during_ suspend.  It is
difficult for the test to tell the difference between a suspend that
was cancelled because the alarm fired too early and a suspend that was

This was cherry-picked from the kernel-next.git akpm branch.

BUG=chromium:239217
TEST=Manual, run suspend/wake testing on DUT, and some test commands on
the wakealarm file:
localhost rtc0 # cat since_epoch
1368751934
localhost rtc0 # echo 1368760000 > wakealarm
localhost rtc0 # cat wakealarm
1368760000
localhost rtc0 # echo +=5 > wakealarm
localhost rtc0 # cat wakealarm
1368760005
localhost rtc0 # echo 0 > wakealarm
localhost rtc0 # cat wakealarm
localhost rtc0 # echo +=5 > wakealarm
-bash: echo: write error: Invalid argument

Signed-off-by: Bernie Thompson <bhthompson@chromium.org>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Cc: Doug Anderson <dianders@chromium.org>
Cc: "Rafael J. Wysocki" <rjw@sisk.pl>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
(cherry picked from commit 19063741561c13df3d3d09138327c7f6f4256ce7)

Change-Id: Ib456d8823f1c95eeae4397fe54325a1e655f8a5b
Reviewed-on: https://gerrit.chromium.org/gerrit/56416
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>