s5p-mfc: encoder VIDIOC_STREAMON wait on HW init
The encoder initialization sequence for s5p-mfc relies on an
initialization job sent to the HW to determine the number of DPB buffers
required for encoding.  Make VIDIOC_STREAMON for the CAPTURE queue wait
until this job is complete, or the subsequent setup of the OUTPUT queue
may race with the HW job and not receive the required DPB count in time.

BUG=chrome-os-partner:18019
TEST=build, run on CrOS snow

Change-Id: Ia99115ddf992233df4335b11c567b03fe38da8e1
Signed-off-by: John Sheu <sheu@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/56507
Reviewed-by: John Sheu <sheu@chromium.org>
Tested-by: John Sheu <sheu@chromium.org>
Commit-Queue: John Sheu <sheu@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59256
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: kiran <kiran@chromium.org>
Tested-by: kiran <kiran@chromium.org>