clk/exynos5420: assign sclk_pixel id to pixel clock divider
sclk_pixel is used to represent pixel clock divider on all exynos
SoCs not as a gate clock. It is queried in driver to pass as the
parent to hdmi clock while switching between parents. A new ID can
be asssigned Pixel gate clock which is currently not in use. Pixel
clock gate is default 'on'.

BUG=chrome-os-partner:19785
TEST=basic hdmi works fine on Snow, smdk5420, peach board.

Change-Id: Ie8e89325b8eb880ef2f49f591a7fdc1ae8d92bb3
Signed-off-by: Rahul Sharma <rahul.sharma@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/56998
Commit-Queue: Shirish S <shirish@chromium.org>
Reviewed-by: Shirish S <shirish@chromium.org>
Tested-by: Shirish S <shirish@chromium.org>