CHROMIUM: i2c: mux: Use subsys_initcall for the i2c-arb-gpio-challenge
It shouldn't be necessary to init the the i2c-arb-gpio-challenge
early.  ...but...

Since many drivers rely on FETs that live behind this arbitrator, they
can't successfully probe until after the arbitrator comes up.  They
ought to handle things properly with EPROBE_DEFER and still work, but
that has some downsides:

1. Those drivers don't come up till later in the boot process.  That
   really not so nice for the LCD--we want that to init early.
2. Some drivers have bugs and don't handle EPROBE_DEFER.  Those
   drivers should be fixed but not all of them have been fixed yet.
   HDMI is one example since DRM doesn't really have good support for
   deferring probes.

Because of the above two reasons, we shuld take this change.  We could
try submitting upstream, but I'm not sure they'll take it.  It's a
pretty small patch to carry locally, though.

We end up using the same init level as the main i2c bus, now.

BUG=chromium:239659
TEST=Boot and see no defer.  AKA: dmesg | grep lcd DOESN'T show:
  [    1.456267] platform platform-lcd.0: Driver platform-lcd requests probe deferral
TEST=Boot and see LCD come up earlier.

Change-Id: I2dc07b1edfff911fcb94a468d058c5d36ea17594
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57007
Reviewed-by: Simon Glass <sjg@chromium.org>