ARM: dts: peach-pit: use high-speed mode on hsi2c4
Communication with the max77802 PMIC appears to be unreliable at 1Mhz
in fast-mode.  Since 1Mhz is the upper limit of fast-mode plus, and it
is not clear that the hsi2c controllers support fast-mode plus, use
high-speed mode on that bus instead.  This also explicity sets the bus
frequency to 1Mhz instead of relying on the default value.

BUG=chrome-os-partner:18993
TEST=Booted on peach-pit 30+ times without i2c communication failures
with the PMIC.

Change-Id: I22024d3fdd67e6ca23e6d87b404375c425a99efe
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57056
Reviewed-by: Doug Anderson <dianders@chromium.org>