CHROMIUM: rtc: cmos: Workaround bios clearing rtc control
The bios may clear the rtc control register when resuming the system. Since the
cmos interrupt handler may now be run before the rtc_cmos is resumed, this can
cause the interrupt handler to ignore an alarm since the alarm bit is not set in
the rtc control register. To work around this, check if the rtc_cmos is
suspended and use the stored value for the rtc control register.

BUG=chromium:242788
TEST=power_Resume -i 50
	Make sure there are no RTC errors

Change-Id: Ib0236ab01b69d96fb673dc29ce2aef7b2fd7f440
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57131
Reviewed-by: Sameer Nanda <snanda@chromium.org>