CHROMIUM: dts: Add smsc,usb3503 to peach_pit device tree
This patch adds the device tree bindings for the new USB3503 HSIC hub
driver to Peach-Pit. It should probably go into the generic file
(instead of rev1), but that counts for most other things in there too,
so let's put it with the other I2C devices for now and wait for someone
to migrate them in bulk.

BUG=chrome-os-partner:19875
TEST=Boot pit rev1, run lsusb, make sure 0424:3503 and 0bda:0129 show up

Change-Id: Id24b80f7d0039e7263216ee20c6d4477eeee22ad
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57160
Reviewed-by: Simon Glass <sjg@chromium.org>