BACKPORT: drm/i915: add support for 5/6 data buffer partitioning on Haswell
Now we compute the results for both 1/2 and 5/6 partitioning and then
use hsw_find_best_result to choose which one to use.

With this patch, Haswell watermarks support should be in good shape.
The only improvement we're missing is the case where the primary plane
is disabled: we always assume it's enabled, so we take it into
consideration when calculating the watermarks.

v2: - Check the latency when finding the best result

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 861f3389c6627460bcd19d1442eb650001f15c9b)

Original-Change-Id: Ia4adcb596cbe0be9c1d3536d471bbcbcc8b8b824
Change-Id: Iafc19762ac82faedf34ca9440819534abafcad85
Reviewed-on: https://gerrit.chromium.org/gerrit/57334
Tested-by: Jim Bride <jim.bride@intel.com>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>