clk: samsung: Fix pll36xx_recalc_rate to handle kdiv properly
The KDIV value is often listed as unsigned but it needs to be treated
as a 16-bit signed value when using it in calculations.  Fix our rate
recalculation to do this correctly.

Before doing this, I tried setting EPLL on exynos5250 to:
  rate, m, p, s, k = 80000000, 107, 2, 4, 43691

This rate is exactly from the table in the exynos5250 user manual.

I read this back as 80750003 with:
  cat /sys/kernel/debug/clk/fin_pll/fout_epll/clk_rate

After this patch, it reads back as 80000003

This patch was posted upstream at:
  https://patchwork.kernel.org/patch/2645361/

BUG=chrome-os-partner:19772
TEST=See above.  I tested on snow and set things manually in U-Boot:
  mw 0x10020130 a06b0204
  mw 0x10020134 0000aaab

Change-Id: I691c24157416922ef726335b1a446bd8ed0a9420
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57233
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Vikas Sajjan <vikas.sajjan@samsung.com>