clk: exynos: Adjust EPLL settings for 49152000 to 49, 3, 3, 9962
There is an issue where the audio code doesn't deal well when we use
(49, 3, 3, 9961).  It's believed that this issue has to do with the
fact that the clk_get_rate() when we use 9961 is just under the
required frequency and code doesn't like that.  Specifically:
  49, 3, 3, 9961 = 49151992 Hz
  49, 3, 3, 9962 = 49152008 Hz

Long term it would be nice to evaluate the code to make it more
robust, but for now let's fix the bug.  See http://crosbug.com/p/19879
for cleaning up the table anyway.

BUG=chrome-os-partner:19772
BUG=chrome-os-partner:19578
TEST=Boot up and play 48 kHz audio

Change-Id: Ife09d9741c163159a4d4349f99a3318192d68420
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57330
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>