drm/exynos: Don't gate register writes on updated
This patch always updates the mixer registers mixer_graph_buffer
instead of gating their value on the updated state variable. This
ensures that we always maintain a consistent state between the
mixer layer and the crtc layer. An example of where this might cause
problems is if crtc updates the dma_addr twice in one vsync. The second
update wouldn't be written to registers like it expects and freeing the
first fb (which should no longer be used) would cause an iommu crash
since it would still be scanned out by mixer.

BUG=None
TEST=Tested on HDMI, no tearing or faults observed

Change-Id: I7f7ced03ca0803b450837b00071b7dabd68674aa
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57506