ARM: exynos: add debug_ll_io_init() call in exynos_init_io()
If the early MMU mapping of the UART happens to get booted out of the
TLB between the start of paging_init() and when we finally re-add the
UART at the very end of s3c_init_cpu(), we'll get a hang at bootup if
we've got early_printk enabled.  Avoid this hang by calling
debug_ll_io_init() early.

Without this patch, you can reliably reproduce a hang when early
printk is enabled by adding flush_tlb_all() at the start of
exynos_init_io().  After this patch the hang goes away.

This is a backport of what I posted upstream at:
  https://patchwork.kernel.org/patch/2673881/

BUG=chrome-os-partner:19950
TEST=Add flush_tlb_all() to the start of exynos_init_io() and things
boot fine.

Change-Id: I09d388c21c619cf196fdda892a2998a857e7ef80
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57558
Reviewed-by: Olof Johansson <olofj@chromium.org>