ARM: gic: add CPU migration support
The git_migrate_target() changes the CPU interface mapping for the
current CPU to redirect SGIs to the given interface, and it also updates
the target CPU for each interrupts to the provided CPU interface if
they were targeting the current interface.

This must be called before gic_cpu_save() so no new SGIs are routed to
the current interface while the already pending ones are preserved and
eventually restored on the new interface via gic_cpu_restore().

BUG=chrome-os-partner:19275
TEST=Booted peach

Change-Id: Id997281f5aa8518cdde3b7f63d86506144e0f0d4
Signed-off-by: Nicolas Pitre <nico@linaro.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57956
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>