BACKPORT: xhci: Don't warn on empty ring for suspended devices.
When a device attached to the roothub is suspended, the endpoint rings
are stopped.  The host may generate a completion event with the
completion code set to 'Stopped' or 'Stopped Invalid' when the ring is
halted.  The current xHCI code prints a warning in that case, which can
be really annoying if the USB device is coming into and out of suspend.

Remove the unnecessary warning.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
Tested-by: Stephen Hemminger <stephen@networkplumber.org>
(cherry picked from commit a83d6755814e4614ba77e15d82796af0f695c6b8)

BUG=chrome-os-partner:19975
TEST=manual: boot slippy device and check for messages of the form:

xhci_hcd 0000:00:14.0: WARN Event TRB for slot 1 ep 6 with no TDs queued?

Change-Id: Ic17e5a82a3c81d94f10ee41918b52a07d9ad04ea
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59075
Reviewed-by: Julius Werner <jwerner@chromium.org>