exynos-gsc: swap cb/cr only for 3 plane formats
The address for cb/cr needs to be swapped for 3 plane formats like
YVU420 and YVU420M. If these address gets swapped for other formats like
NV21, it results in passing a NULL dma address to gscalar (which will
result in a PAGE FAULT if sysmmu is enabled).

E.g. For NV21, the dma_address are (Y, CbCr, 0) and we swap them (Y, 0,
CbCr) which is incorrect.

BUG=chrome-os-partner:19007
TEST=Run gscalar test application (gsc-m2m-test) with NV21/NV61 format
as input. This should run successfully without PAGEFAULT.

Change-Id: Ib1fe06dc6b90bc68d57d86c9e709742a3edb636e
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Shirish S <s.shirish@samsung.com>
Signed-off-by: Arun Mankuzhi <arun.m@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/57715
Reviewed-by: Arun Mankuzhi <arun@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Tested-by: Shirish S <shirish@chromium.org>
Commit-Queue: Shirish S <shirish@chromium.org>