CHROMIUM: config: Make s3c-rtc a module.
By making s3c-rtc a module and the primary RTC (from Exynos PMICs) for
exynos based designs built-in to the kernel it should guarantee the
primary RTC is /dev/rtc0 which is battery backed so it will retain
time in S5.  This device is then used by hctosys in the kernel as well
as the tlsdated daemon in userspace.  The s3c-rtc is used for autotest
purposes as it doesn't have the same problems delivering interrupts
that the PMIC based RTCs do.

Signed-off-by: Todd Broch <tbroch@chromium.org>

CQ-DEPEND=CL:56667
BUG=chromium:221857
TEST=manual
On daisy & spring:
     $ ls lib/modules/3.4.0/kernel/drivers/rtc/rtc-s3c.ko
     lib/modules/3.4.0/kernel/drivers/rtc/rtc-s3c.ko

     # for daisy
     $ basename $(readlink /sys/class/rtc/rtc0/device/driver)
     rtc-max77686

     # for spring
     $ basename $(readlink /sys/class/rtc/rtc0/device/driver)
     s5m-rtc

     $ basename $(readlink /sys/class/rtc/rtc1/device/driver)
     s3c-rtc

     $ echo "+10s" > /sys/class/rtc/rtc1/wakealarm && echo mem > /sys/power/state

     <system wakes in ~10sec>

Change-Id: I2dcf0c2fa660fbdc887011a59948a675e658d82d
Reviewed-on: https://gerrit.chromium.org/gerrit/57875
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Todd Broch <tbroch@chromium.org>