CHROMIUM: iwl7000: add new driver
Add the "iwl7000" driver, supporting 7000 and 3000 series Intel
wireless devices. This driver has its own copy of mac80211,
properly modified to be concurrently usable with the built-in
mac80211, and thus behaves more like a full-MAC driver.

This is a preliminary version of the driver, based on a tested
version that might still be missing features.

BUG=chrome-os-partner:20800
TEST=new iwl7000 driver passes (most) WiFi tests

Change-Id: I76d0c60e7531ba7e29f0b2860aa5ba93f7fd6126
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/59784
Reviewed-by: Paul Stewart <pstew@chromium.org>