UPSTREAM: USB: cdc-acm: remove unneeded spin_lock_irqsave/restore on write path
When writing data we were:
	lock
	do some work
	unlock
	call function
		lock
		do some work
		unlock
		return
	return

It turns out, that "function" was only ever called in the one place, so
instead of locking/unlocking for no good reason, just inline the
function and only grab the lock once.

This has sped up the pathological case of sending 1 byte packets to a
loop-back cdc-acm device from 49600 bytes per second to 50100 bytes a
second on my workstation.  A tiny increase yes, but noticable, and now
the spinlock isn't the hottest thing on the perf graph anymore.  Yes, we
are still waiting for the hardware for the most part, but getting rid of
a spinlock_irq_save() call for every packet is still a good thing.

And we end up deleting lines of code, always a win overall.

This was found by using a Teensy 3.0 device and the test program and
firmware located at:
	http://www.pjrc.com/teensy/benchmark_usb_serial_receive.html

Reported-by: Paul Stoffregen <paul@pjrc.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

BUG=chromium:345086
TEST=Tested the following:
     - emerge-daisy_spring chromeos-kernel-next
     - Manually tested cellular support on a DUT with alt3100 and y3300 modems
     - On the same devices, ran network_3GIdentifiers, network_3GModemPresent,
       network_3GSmokeTest

(cherry picked from commit 71c2fb03fc5800f24b66b730cbb9fddc22df40d3)
Signed-off-by: Prathmesh Prabhu <pprabhu@chromium.org>
Change-Id: Ifcfc6db159eb83e42668f4b367061e2231c3eb41
Reviewed-on: https://chromium-review.googlesource.com/187303
Reviewed-by: Ben Chan <benchan@chromium.org>