UPSTREAM: mmc: core: Enable power_off_notify for eMMC shutdown sequence
In suspend mode it is important to save power. If the host is able to
cut buth vcc and vccq, the MMC_CAP2_POWEROFF_NOTIFY shall be set. It
will mean the card will be completely powered down at suspend and the
power off notification cmd will be sent prior power down.

It seems common not being able to cut both vcc and vccq for a host. In
this situation we issue the sleep cmd in favor of the power off
notification cmd, to save more power.

While maintainng the above policy, we also want to make use of the
power off notification in the shutdown sequence, even in the case were
the host has not set MMC_CAP2_POWEROFF_NOTIFY, since we know vcc and
vccq will regardless be cut.

We accomplish this by always enabling the power off notification byte
in the EXT_CSD and issue the power off notification when either
MMC_CAP2_POWEROFF_NOTIFY is set or we are executing a shutdown.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Chris Ball <cjb@laptop.org>
(cherry picked from commit 43235679341510f58a86eb75ab23cbae4bdc59b8)

BUG=chromium:383481
TEST=run bvt-inline and bvt-cq suite on spring

Change-Id: I56a590c9cb1f5c5f59bb19513750579106789d5f
Reviewed-on: https://chromium-review.googlesource.com/212100
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>