CHROMIUM: x86, relocs: move __vvar_page from S_ABS to S_REL
The __vvar_page relocation should actually be listed in S_REL instead
of S_ABS.

Additional details from
https://prodkernel-review.git.corp.google.com/#/c/32286/

The symbol __vvar_page is the kernel virtual address of the page
that is exported to user space in the VVAR_PAGE entry of the fixmap
and which contains the data needed by the functions implemented
in the VDSO.

The linker marks this symbol as being absolute because of the way
in which it is defined in vmlinux.lds.S but it needs to be relocated
when ASLR is in use.

relocs.c uses regular expression matching of symbol names to handle
symbols like this that are incorrectly categorized by the linker.
Unfortunately __vvar_page was added to the wrong regexp resulting
in relocations referencing the symbol being silently ignored when
they should actually have been applied.

Tested: booted kernel with ASLR enabled and verified that system
        calls that use the kernel VDSO - time(), gettimeofday()
        and clock_gettime() return appropriate values

BUG=chromium:230131
TEST=link build, manual testing

Change-Id: Ibcd931f66bcab3adc55370855c255573dda2157d
(pending in tip http://git.kernel.org/cgit/linux/kernel/git/tip/tip.git/commit/?id=b1983b0a7578de09211a696802edab83fd253303)
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58483
Reviewed-by: Aaron Durbin <adurbin@chromium.org>