clk: samsung: specify both lock and control registers for pll35xx/pll36xx
The PLL driver currently assume that the control registers are at an
offset of 0x100 from the lock register for pll35xx and pll36xx PLLs.
While this is true for Exynos4 and Exynos5250, it is not for Exynos5420.
This modifies the PLL driver so that the base address of both the lock
and control registers are specified when registering a pll35xx/pll35xx
PLL.

BUG=chrome-os-partner:20104
TEST=Boots on Snow and Pit; audio still works on both.

Change-Id: Ie9433ff577ee569c4a5a34b1f588b78a9a11afa8
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58411
Reviewed-by: Doug Anderson <dianders@chromium.org>