usb: dwc3-exynos: Do not cut VBUS power on suspend
The dwc3-exynos host controller platform driver uses a custom GPIO pin
to control its external port's VBUS power on some boards. The current
code disables this GPIO on system suspend.

This is inconsistent with how the rest of Linux behaves -- other
platform drivers don't remove USB power unconditionally on suspend. This
patch removes the responsible line so that affected dwc3-exynos boards
support charging of USB devices while suspended. The code to reactivate
the GPIO on resume stays in place as a safeguard, just in case some
other agent (firmware?) messed with it in-between.

BUG=chrome-os-partner:18005
TEST=Plug a non-Apple phone into the USB3 port on a Snow. Suspend the
system, notice how the phone is still charging.

Change-Id: Ie4df264250fa0c85932b590e64cf5e781fc85e71
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58424
Reviewed-by: Simon Glass <sjg@chromium.org>