CHROMIUM: arm: exynos5420: Add clock muxing setup; reparent SPI
We somehow need to setup clock muxing and would rather not rely on
U-Boot setting it up properly for us (it makes us fragile against
U-Boot changes).

The setup of clock muxing is a mix of the design of the board and the
way that the software is currently setup.  Because it's not purely
defined by the board, it can't live in the device tree.  That means
that (at the moment) the best place for it is in code.

At the moment we just use this muxing infrastructure to setup SPI
clocks.  We should go through and add other muxing as needed (or
perhaps fill in the whole table?)

This is not a great solution but is a stopgap until someone
(upstream?) comes up with a better solution.  Stopgap solution
suggested by Olof Johansson.

BUG=chrome-os-partner:20013
TEST=Run 'find /sys/kernel/debug/clk -name mout_spi0'
/sys/kernel/debug/clk/fin_pll/fout_mpll/sclk_mpll/mout_spi0

Change-Id: Ic36bef6fa991aed7138d69b5216aa8fcd59997c9
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58454
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>