CHROMIUM: drm/exynos: Force vblanks on before waiting in mode_set_base
When a page_flip happens after a DPMS OFF, the vblank interrupts for
that flip never arrive untile the display turns back on. If the next
thing we do is mode_set_base, the waiting routing in there will
timeout waiting for events. Worse, the buffer will remain there and
we will destroy it, leading to an mmu page fault.

To fix that, we force DPMS ON and get vblank before mode_set_base to
ensure that vblanks will arrive when we need them.

BUG=chromium:249547
TEST=by hand: while poster circle is running, do {suspend/resume, xset dpms force off, powerd_suspend} a couple times

Change-Id: I7a254ad10dce71f893e4c67a48d1a87dd163d853
Reviewed-on: https://gerrit.chromium.org/gerrit/58616
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Stéphane Marchesin <marcheu@chromium.org>