ARM: dts: Add mask-tpm-reset to the device tree
The mask-tpm-reset GPIO is used by the kernel to prevent the TPM from
being reset across sleep/wake.  If we don't set it to anything then
the TPM will be reset.  U-Boot will soon be detecting this as invalid
and will reset the system on resume time.  This GPIO can always be low
and not hurt anything.  It will get pulled back high again during a
normal warm reset when it will default back to an input.

To properly preserve the TPM state across suspend/resume and to make
the future version of U-Boot happy, properly set the GPIO to mask the
reset to the TPM.

BUG=chrome-os-partner:20172
TEST=With new U-Boot 'suspend_stress_test' no longer reboots the
system.

Change-Id: Ie01e1578e136809577127695504109625c5a98f6
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58737
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>