cpufreq: exynos: ensure the core voltage is always right at suspend
On the resume path, the bootloader is restoring the max frequency
(1700Mhz) without touching the PMIC. So we need to ensure that the PMIC
core voltage is high enough to support the full frequency range.
The cpufreq code is trying to do that by having a "locking_frequency"
which is forced when the CPU is being suspended.
In order to guarantee the right voltage setting :
- fix the "locking_frequency" code to be able to set the frequency
  regardless of the current governor will.
- ensure the "locking_frequency" is always the maximum frequency
  (triggering the maximum core voltage) and not the current value set at
  startup (which might be different, e.g. low battery throttling).

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:19895
TEST=on Spring, run the stress script provided in the issue tracker,
and do suspend/resume cycles.
Also trace all voltage and frequency changes and check we are getting
the expected values.

Change-Id: I2e9c3eb0e69ad16164f115e22637c0457bbb121c
Reviewed-on: https://gerrit.chromium.org/gerrit/58925
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>