Revert "CHROMIUM: config: Add new usb3503 driver to chromeos-exynos5"
This reverts commit 4043b2ee0452f2e35379eed46bfb8e3ba5020010.

The usb3503 I2C driver has insurmountable problems to correctly bring up
the HSIC hub when it had already been active and configured before (e.g.
by firmware). Since we want a kernel that can initialize its devices to
a good state in all situations, this is not acceptable and we need to
fall back to a closer integration with the PHY driver.

BUG=chrome-os-partner:20205
TEST=None

Change-Id: I5fe464a99dc9be9fd9922e0f9b73b046e453ae80
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58960
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>