UPSTREAM: perf/x86/intel: Add Haswell PEBS record support
Add support for the Haswell extended (fmt2) PEBS format.

It has a superset of the nhm (fmt1) PEBS fields, but has a
longer record so we need to adjust the code paths.

The main advantage is the new "EventingRip" support which
directly gives the instruction, not off-by-one instruction. So
with precise == 2 we use that directly and don't try to use LBRs
and walking basic blocks. This lowers the overhead of using
precise significantly.

Some other features are added in later patches.

BUG=chromium:371570
TEST=`emerge-link chromeos-kernel-next`

Reviewed-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Cc: Andi Kleen <ak@linux.jf.intel.com>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Arnaldo Carvalho de Melo <acme@redhat.com>
Link: http://lkml.kernel.org/r/1371515812-9646-2-git-send-email-andi@firstfloor.org
Signed-off-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit 130768b8c93cd8d21390a136ec8cef417153ca14)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>

Change-Id: Id4089ab5563cb5bcb319e171db3d2c59ad32ff65
Reviewed-on: https://chromium-review.googlesource.com/199991
Reviewed-by: Simon Que <sque@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>