Revert "CHROMIUM: exynos: reset external HSIC hub"
This reverts commit 079da6cbc07b3b11144232bcd06adcd57f889492.

I'm really not sure why this mechanism doesn't work with a
firmware-initialized HSIC hub, since that's how we did it in 3.4. Maybe
something about regulator_force_disable() changed and doesn't really do
anything when the regulator hadn't been "activated" by the kernel
before?

Anyway, I think the regulator framework isn't such a great way to model
this in the first place (it's just a reset signal, not a power line
after all). We have another patch that seems to work where this one
doesn't, so let's take this out and go that route instead.

BUG=chrome-os-partner:20205
TEST=None

Change-Id: I9c38784a03dbdbf842edcb83f6908e62229c8785
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/58962
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>